only require structopt for CLI
