+++ Cargo.toml
@@ -25,9 +25,13 @@ exclude = [
 travis-ci = { repository = "rust-fuzz/honggfuzz-rs", branch = "master" }
 maintenance = { status = "actively-developed" }
 
+[features]
+
+cli = ["structopt"]
+
 [dependencies]
 arbitrary = "1"
-structopt = "0.3"
+structopt = { version = "0.3", optional = true }
 
 [dev-dependencies]
 rand = "0.8"
@@ -38,3 +42,8 @@ lazy_static = "1.4"
 
 [target.'cfg(fuzzing_debug)'.dependencies]
 memmap = "0.7"
+
+[[bin]]
+name = "cargo-hfuzz"
+# requires the `cli` feature, but if it is forgotten, we want to fail compilation instead of silently skipping this binary
+# we use conditionnal compilation to throw the error
